requirements.txt
flask
PyYAML
boto3
sqlite-web